+++ src/js/actions/AppActions.js
@@ -2,19 +2,18 @@ var AppDispatcher = require('../dispatcher/AppDispatcher');
 var AppConstants = require('../constants/AppConstants');
 
 var AppActions = {
-  searchMovies: function(movie){
-    AppDispatcher.handleViewAction({
-      actionType: AppConstants.SEARCH_MOVIES,
-      movie: movie
-    });
-  },
-  receiveMovieResults: function(movies){
-    console.log(movies);
-    AppDispatcher.handleViewAction({
-      actionType: AppConstants.RECEIVE_MOVIE_RESULTS,
-      movies: movies
-    });
-  }
+	searchMovies: function(movie){
+		AppDispatcher.handleViewAction({
+			actionType: AppConstants.SEARCH_MOVIES,
+			movie: movie
+		});
+	},
+	receiveMovieResults: function(movies){
+		AppDispatcher.handleViewAction({
+			actionType: AppConstants.RECEIVE_MOVIE_RESULTS,
+			movies: movies
+		});
+	}
 }
 
 module.exports = AppActions;
