+++ src/js/constants/AppConstants.js
@@ -1,4 +1,4 @@
 module.exports = {
-  SEARCH_MOVIES: 'SEARCH_MOVIES',
-  RECEIVE_MOVIE_RESULTS: 'RECEIVE_MOVIE_RESULTS'
+	SEARCH_MOVIES: 'SEARCH_MOVIES',
+	RECEIVE_MOVIE_RESULTS: 'RECEIVE_MOVIE_RESULTS'
 }
