+++ src/js/main.js
@@ -1,9 +1,9 @@
-var App = require('./components/App.jsx');
+var App = require('./components/App');
 var React = require('react');
 var ReactDOM = require('react-dom');
-var AppAPI = require('./utils/AppAPI');
+var AppAPI = require('./utils/appAPI.js');
 
 ReactDOM.render(
-  <App />,
-  document.getElementById('app')
+	<App />,
+	document.getElementById('app')
 );
